game history showPreviousMove now revertingAlteredColors

+++ Geminy Cricket/RFHViewController.m
@@ -651,7 +651,7 @@
         label.textColor = [UIColor colorWithRed:.22 green:.8 blue:.33 alpha:1.0];
         //label.font = [label.font fontWithSize:25];
         label.font = [UIFont fontWithName:@"Zapfino" size:15];
-        completedGame.outcome = [NSString stringWithFormat:@"%@ (%lu - %lu)", labelText, humanTotal, robotTotal];
+        completedGame.outcome = [NSString stringWithFormat:@"%@ (%lu - %lu)", labelText, (unsigned long)humanTotal, (unsigned long)robotTotal];
         completedGame.humanVictory = YES;
         [self.view addSubview:label];
     } else {
@@ -666,7 +666,7 @@
         label.text = labelText;
         label.textColor = [UIColor colorWithRed:.8 green:.22 blue:.1 alpha:1.0];
         label.font = [UIFont fontWithName:@"Zapfino" size:15];
-        completedGame.outcome = [NSString stringWithFormat:@"%@ (%lu - %lu)", labelText, humanTotal, robotTotal];
+        completedGame.outcome = [NSString stringWithFormat:@"%@ (%lu - %lu)", labelText, (unsigned long)humanTotal, (unsigned long)robotTotal];
         [self.view addSubview:label];
     }
